Just add the methods from the helper file to the test file.
