+++ activesupport/test/dependencies_test.rb
@@ -1,7 +1,6 @@
 require 'abstract_unit'
 require 'pp'
 require 'active_support/dependencies'
-require 'dependencies_test_helpers'
 
 module ModuleWithMissing
   mattr_accessor :missing_count
@@ -19,8 +18,6 @@ class DependenciesTest < ActiveSupport::TestCase
   def teardown
     ActiveSupport::Dependencies.unload!
   end
-
-  include DependenciesTestHelpers
 
   def test_depend_on_path
     skip "LoadError#path does not exist" if RUBY_VERSION < '2.0.0'
@@ -778,6 +775,25 @@ class DependenciesTest < ActiveSupport::TestCase
   end
 
 private
+  def with_loading(*from)
+    old_mechanism = ENV.delete('NO_RELOAD')
+    this_dir = File.dirname(__FILE__)
+    parent_dir = File.dirname(this_dir)
+    path_copy = $LOAD_PATH.dup
+    $LOAD_PATH.unshift(parent_dir) unless $LOAD_PATH.include?(parent_dir)
+    prior_autoload_paths = ActiveSupport::Dependencies.autoload_paths
+    ActiveSupport::Dependencies.autoload_paths = from.map { |f| File.join(this_dir, f) }
+    yield
+  ensure
+    $LOAD_PATH.replace(path_copy) if path_copy
+    ActiveSupport::Dependencies.autoload_paths = prior_autoload_paths
+    ENV['NO_RELOAD'] = old_mechanism
+  end
+
+  def with_autoloading_fixtures(&block)
+    with_loading 'autoloading_fixtures', &block
+  end
+
   def const_scope(const_name)
     with_autoloading_fixtures do
       yield
